// types/Tools.ts
interface Tool {
  addedDate: Date;
  category: string;
  imageUrl: string[];
  rentCount: number;
  sizes: string;
  toolsName: string;
  toolsRealPrice: string;
  updatedDate: Date;
  description: string;
  id: string;
  toolsAvailability: boolean;
  toolsCondition: string;
  toolsRentPrice: string;
  sizesName: string;
  updatedAt: { seconds: number };
  createdAt: { seconds: number };
  updatedBy: string;
  createdBy: string;
  rentProfit: string;
}